Restore and organize date-related "should" statements in all tests.
In order to reduce confusion about what has and has not changed in
the test suite, this commit restores the "should" statements in each
test and organizes them to follow a common, consistent pattern.

To summarize, the only changes that have been made to the test suite are:

* The refactoring of the nock data into its own file (util/nocks.js)
* The organization of "should" statements for consistency across tests
* The addition of tests for the new time-based preference features

The first increases the DRYness of the test suite, the 2nd keeps things
consistent, and the 3rd exercises the changes made to prefer recency in
fetched results.